% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGen.R
\name{netgen}
\alias{netgen}
\title{netgen}
\usage{
netgen(net_size = 50, ave_module_size = 10, min_module_size = 1,
  min_submod_size = 1, net_type = c("mixed", "random", "scalefree",
  "nested", "bn", "br", "tt-bn-r", "tt-bn-bn"), ave_degree = 10,
  rewire_probs = c(0.3, 0), mod_probs = 0, verbose = FALSE)
}
\arguments{
\item{net_size}{network size (number of nodes)}

\item{ave_module_size}{average module size}

\item{min_module_size}{cutoff for the minimum modules size}

\item{min_submod_size}{cutoff for submodules, used only for bipartite and tripartite networks}

\item{net_type}{network type, see details}

\item{ave_degree}{average degree of connection}

\item{rewire_probs}{global and local  network rewiring probabilities}

\item{mod_probs}{module probabilities for first 7 types,
used for constructing mixed networks}

\item{verbose}{logical, default TRUE. Should a message report summary statistics?}
}
\value{
an \code{igraph} object
}
\description{
Randomly generate a wide range of interaction networks
}
\details{
network type is one of
\itemize{
\item mixed
\item random
\item scalefree
\item nested
\item bi-partite nested
\item bi-partite random
\item tri-trophic bipartite nested-random. (Can use short-hand "ttbnr")
\item tri-trophic bipartite nested-bipartite nested (Can use short-hand "ttbnbn")
}

NOTE: Function arguments have changed from those netgen 0.1.1 to be more intelligible.
To restore the original function api on code that depends on the old version, you
can simply add:

\code{netgen <- EcoNetGen:::netgen_v1}

to the top of your code after running \code{library(EcoNetGen)}.
}
\examples{
library(EcoNetGen)
\donttest{
net <- netgen()
adj_plot(net)
}
}
