% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGen.R
\name{netgen_v1}
\alias{netgen_v1}
\title{netgen_v1}
\usage{
netgen_v1(n_modav = c(50, 10), cutoffs = c(3, 0), net_type = 1,
  net_degree = 10, net_rewire = c(0.3, 0), mod_probs = c(0.2, 0.2, 0.2,
  0.2, 0.2, 0, 0), verbose = FALSE)
}
\arguments{
\item{n_modav}{network size and average module size (integer vector, length 2)}

\item{cutoffs}{module and submodule minimum sizes (integer vector, length 2).
(submodules are used only for bipartite and tripartite networks)}

\item{net_type}{integer indicating type, see details}

\item{net_degree}{average degree of connection}

\item{net_rewire}{global and local  network rewiring probabilities}

\item{mod_probs}{module probabilities for types 1 to 51,
used for constructing mixed networks, net_type = 0}

\item{verbose}{logical, default TRUE. Should a message report summary statistics?}
}
\value{
an \code{igraph} object
}
\description{
netgen function
}
\details{
network type
\itemize{
\item 0 = mixed
\item 1 = random
\item 2 = scalefree
\item 3 = nested
\item 41 = bi-partite nested
\item 42 = bi-partite random
\item 51 = tri-trophic bipartite nested-random "ttbnr"
\item 52 = tri-trophic bipartite nested-bipartite nested "ttbnbn"
}
}
