% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetGen.R
\name{netgen}
\alias{netgen}
\title{netgen}
\usage{
netgen(net_size = 50, average_module = 10, min_module = 0,
  min_submodule = 0, net_type = c("mixed", "random", "scalefree", "nested",
  "bi-partite nested", "bi-partite random", "tt-bn-r", "tt-bn-bn"),
  net_degree = 10, net_rewire = c(0.3, 0), mod_probs = 0,
  verbose = TRUE)
}
\arguments{
\item{net_size}{network size (number of nodes)}

\item{average_module}{average module size}

\item{min_module}{cutoff for the minimum modules size}

\item{min_submodule}{cutoff for submodules, used only for bipartite and tripartite networks}

\item{net_type}{network type, see details}

\item{net_degree}{average degree of connection}

\item{net_rewire}{global and local  network rewiring probabilities}

\item{mod_probs}{module probabilities for first 7 types,
used for constructing mixed networks}

\item{verbose}{logical, default TRUE. Should a message report summary statistics?}
}
\value{
an \code{igraph} object
}
\description{
Randomly generate a wide range of interaction networks
}
\details{
network type is one of
\itemize{
\item mixed
\item random
\item scalefree
\item nested
\item bi-partite nested
\item bi-partite random
\item "ttbnr": tri-trophic bipartite nested-random
\item "ttbnbn": tri-trophic bipartite nested-bipartite nested
}

NOTE: Function arguments have changed from those netgen 0.1.1 to be more intelligible.
To restore the original function api on code that depends on the old version, you
can simply add:

\code{netgen <- EcoNetGen:::netgen_v1}

to the top of your code after running \code{library(EcoNetGen)}.
}
\examples{
\donttest{
net <- netgen()
adj_plot(net)
}
}
